Updated to new cache loader pattern.
